refactor(control): split reload subscription
