+++ crates/honk-core/src/control/mod.rs
@@ -24,7 +24,6 @@ mod probers;
 pub mod quic;
 pub(crate) mod reload;
 mod reload_policy;
-mod reload_subscription;
 #[cfg(test)]
 mod reload_tests;
 mod resource_budget;
