+++ crates/honk-core/src/control/reload/subscription.rs
@@ -7,7 +7,7 @@ use super::*;
 /// merged node set. Nodes from other subscriptions and static config nodes
 /// are untouched. Re-merging the same subscription is idempotent — nodes
 /// are replaced, never duplicated.
-pub(super) fn config_with_subscription_nodes(
+pub(in crate::control) fn config_with_subscription_nodes(
     current: &Config,
     subscription_id: uuid::Uuid,
     nodes: Vec<Node>,
@@ -29,4 +29,23 @@ pub(super) fn config_with_subscription_nodes(
         &config.subscriptions,
     );
     config
+}
+
+impl ControlPlane {
+    /// Merge freshly fetched subscription nodes into the running config,
+    /// replacing the previous node set of `subscription_id`, and run the
+    /// shared rebuild pipeline.
+    ///
+    /// Production callers go through `ControlCommand::MergeSubscription` on
+    /// the command channel (which keeps merges serialized against SIGHUP
+    /// reloads); this public wrapper exists so integration tests can drive a
+    /// merge without binding the TPROXY accept loop.
+    pub async fn merge_subscription_nodes(&self, subscription_id: uuid::Uuid, nodes: Vec<Node>) {
+        let new_config = {
+            let current = self.config.read().await;
+            config_with_subscription_nodes(&current, subscription_id, nodes)
+        };
+        let drain = DrainTracker::new();
+        self.apply_runtime_config(new_config, &drain).await;
+    }
 }
